moved vpn types out of shared
